# frontend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Minilog API",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/v1/imports/piyolog/preview": {
      "post": {
        "tags": [
          "PiyoLog import"
        ],
        "summary": "Preview Import",
        "operationId": "preview_import_api_v1_imports_piyolog_preview_post",
        "parameters": [
          {
            "name": "X-CSRF-Token",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Csrf-Token"
            }
          },
          {
            "name": "minilog_session",
            "in": "cookie",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Minilog Session"
            }
          },
          {
            "name": "minilog_csrf",
            "in": "cookie",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Minilog Csrf"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_preview_import_api_v1_imports_piyolog_preview_post"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PiyoLogPreview"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/imports/piyolog/confirm": {
      "post": {
        "tags": [
          "PiyoLog import"
        ],
        "summary": "Confirm Import",
        "operationId": "confirm_import_api_v1_imports_piyolog_confirm_post",
        "parameters": [
          {
            "name": "X-CSRF-Token",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Csrf-Token"
            }
          },
          {
            "name": "minilog_session",
            "in": "cookie",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Minilog Session"
            }
          },
          {
            "name": "minilog_csrf",
            "in": "cookie",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Minilog Csrf"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_confirm_import_api_v1_imports_piyolog_confirm_post"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ImportBatchOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/imports/piyolog": {
      "get": {
        "tags": [
          "PiyoLog import"
        ],
        "summary": "List Imports",
        "operationId": "list_imports_api_v1_imports_piyolog_get",
        "parameters": [
          {
            "name": "minilog_session",
            "in": "cookie",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Minilog Session"
            }
          },
          {
            "name": "minilog_csrf",
            "in": "cookie",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Minilog Csrf"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ImportBatchOut"
                  },
                  "title": "Response List Imports Api V1 Imports Piyolog Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/imports/piyolog/daily-notes": {
      "get": {
        "tags": [
          "PiyoLog import"
        ],
        "summary": "List Imported Daily Notes",
        "operationId": "list_imported_daily_notes_api_v1_imports_piyolog_daily_notes_get",
        "parameters": [
          {
            "name": "baby_id",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Baby Id"
            }
          },
          {
            "name": "minilog_session",
            "in": "cookie",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Minilog Session"
            }
          },
          {
            "name": "minilog_csrf",
            "in": "cookie",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Minilog Csrf"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ImportedDailyNoteOut"
                  },
                  "title": "Response List Imported Daily Notes Api V1 Imports Piyolog Daily Notes Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/imports/piyolog/{batch_id}/source": {
      "delete": {
        "tags": [
          "PiyoLog import"
        ],
        "summary": "Delete Import Source",
        "operationId": "delete_import_source_api_v1_imports_piyolog__batch_id__source_delete",
        "parameters": [
          {
            "name": "batch_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Batch Id"
            }
          },
          {
            "name": "X-CSRF-Token",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Csrf-Token"
            }
          },
          {
            "name": "minilog_session",
            "in": "cookie",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Minilog Session"
            }
          },
          {
            "name": "minilog_csrf",
            "in": "cookie",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Minilog Csrf"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "Body_confirm_import_api_v1_imports_piyolog_confirm_post": {
        "properties": {
          "baby_id": {
            "type": "string",
            "format": "uuid",
            "title": "Baby Id"
          },
          "source_time_zone": {
            "type": "string",
            "title": "Source Time Zone"
          },
          "file": {
            "type": "string",
            "contentMediaType": "application/octet-stream",
            "title": "File"
          },
          "retain_source": {
            "type": "boolean",
            "title": "Retain Source",
            "default": true
          },
          "replace_modified": {
            "type": "boolean",
            "title": "Replace Modified",
            "default": false
          }
        },
        "type": "object",
        "required": [
          "baby_id",
          "source_time_zone",
          "file"
        ],
        "title": "Body_confirm_import_api_v1_imports_piyolog_confirm_post"
      },
      "Body_preview_import_api_v1_imports_piyolog_preview_post": {
        "properties": {
          "baby_id": {
            "type": "string",
            "format": "uuid",
            "title": "Baby Id"
          },
          "source_time_zone": {
            "type": "string",
            "title": "Source Time Zone"
          },
          "file": {
            "type": "string",
            "contentMediaType": "application/octet-stream",
            "title": "File"
          }
        },
        "type": "object",
        "required": [
          "baby_id",
          "source_time_zone",
          "file"
        ],
        "title": "Body_preview_import_api_v1_imports_piyolog_preview_post"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ImportBatchOut": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "baby_id": {
            "type": "string",
            "format": "uuid",
            "title": "Baby Id"
          },
          "source_hash": {
            "type": "string",
            "title": "Source Hash"
          },
          "detected_locale": {
            "type": "string",
            "title": "Detected Locale"
          },
          "date_from": {
            "anyOf": [
              {
                "type": "string",
                "format": "date"
              },
              {
                "type": "null"
              }
            ],
            "title": "Date From"
          },
          "date_to": {
            "anyOf": [
              {
                "type": "string",
                "format": "date"
              },
              {
                "type": "null"
              }
            ],
            "title": "Date To"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "counts": {
            "additionalProperties": {
              "type": "integer"
            },
            "type": "object",
            "title": "Counts"
          },
          "source_retained": {
            "type": "boolean",
            "title": "Source Retained"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "baby_id",
          "source_hash",
          "detected_locale",
          "date_from",
          "date_to",
          "status",
          "counts",
          "source_retained",
          "created_at"
        ],
        "title": "ImportBatchOut"
      },
      "ImportedDailyNoteOut": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "baby_id": {
            "type": "string",
            "format": "uuid",
            "title": "Baby Id"
          },
          "local_date": {
            "type": "string",
            "format": "date",
            "title": "Local Date"
          },
          "body": {
            "type": "string",
            "title": "Body"
          },
          "source_author_text": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Source Author Text"
          },
          "source_line": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Source Line"
          }
        },
        "type": "object",
        "required": [
          "id",
          "baby_id",
          "local_date",
          "body",
          "source_author_text",
          "source_line"
        ],
        "title": "ImportedDailyNoteOut"
      },
...
      "PiyoLogPreview": {
        "properties": {
          "source_hash": {
            "type": "string",
            "title": "Source Hash"
          },
          "duplicate_import_id": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ],
            "title": "Duplicate Import Id"
          },
          "detected_locale": {
            "type": "string",
            "enum": [
              "en",
              "ja"
            ],
            "title": "Detected Locale"
          },
          "date_from": {
            "anyOf": [
              {
                "type": "string",
                "format": "date"
              },
              {
                "type": "null"
              }
            ],
            "title": "Date From"
          },
          "date_to": {
            "anyOf": [
              {
                "type": "string",
                "format": "date"
              },
              {
                "type": "null"
              }
            ],
            "title": "Date To"
          },
          "counts": {
            "additionalProperties": {
              "type": "integer"
            },
            "type": "object",
            "title": "Counts"
          },
          "conflicts": {
            "additionalProperties": {
              "type": "integer"
            },
            "type": "object",
            "title": "Conflicts"
          },
          "unknown_lines": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "type": "array",
            "title": "Unknown Lines"
          },
          "warnings": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Warnings"
          }
        },
        "type": "object",
        "required": [
          "source_hash",
          "detected_locale",
          "counts",
          "unknown_lines",
          "warnings"
        ],
        "title": "PiyoLogPreview"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}